// src/app/modules/hcm/shared/models/laborRelationship/accountforpayrolldata.ts
export class AccountforPayrollData {

    idLaborRelationship: number = -1;
    accountforPayrollDataId: number = -1;
    bankId: number = -1;
    bank: string = "";
    typeAccountId: number = -1;
    typeAccount: string = "";
    accountNumber: string = "";
    indActive: boolean = false;
    createdById: number = -1;
    modifiedById: number = -1;
    
}