// src/app/modules/hcm/shared/models/masters/bank-account-type.ts
export class BankAccountType {
    id: number = -1;
    name: string = "";
    active: number = -1;
}

export class BankAccountTypeFMS {
    id: number
    name: string
    active: boolean
}

export class BankAccountTypeFMSFilter {
    id: number = -1
    name: string = ""
    active: number = -1
}